Bug 846909 - insert our testing framework at run time.
Bug 846909 - Change the initialisation of contentScript and where we inject.

Rewrite the PerfHelper injection
