+++ apps/settings/test/performance/rendering_wifi_list_test.js
@@ -23,7 +23,12 @@ marionette(mozTestInfo.appPath + ' >', function() {
     // It affects the first run otherwise
     this.timeout(500000);
     client.setScriptTimeout(50000);
+
     MarionetteHelper.unlockScreen(client);
+
+    // inject performance helper
+    client.contentScript.inject(
+      GAIA_DIR + '/tests/performance/performance_testing_helper.js');
   });
 
   test('rendering WiFi list >', function() {
