+++ tests/performance/startup_events_test.js
@@ -44,6 +44,10 @@ marionette('startup event test > ' + mozTestInfo.appPath + ' >', function() {
     client.setScriptTimeout(50000);
 
     MarionetteHelper.unlockScreen(client);
+
+    // inject performance helper
+    client.contentScript.inject(
+      GAIA_DIR + '/tests/performance/performance_testing_helper.js');
   });
 
   test('startup >', function() {
